Fix https://github.com/beyond-all-reason/RecoilEngine/issues/2659
The precision of the select() based sleep in UDPListener::Update(int loopSleepTime) is actually junk. It jumps from values as low as 10 to values as high as 2700 microseconds, with the target being 1 ms.
Nonetheless, the issue caused 0.67x replay speed was actually elsewhere. Funny thing is, that [it was discovered during a similar investigation](https://discord.com/channels/549281623154229250/549282166543089674/1322572999919931493), but the results of that investigation never made it to the commit.